US Senate committee to look at 5G vendor security
US Senate committee to look at 5G vendor security

The US Senate Committee on Commerce, Science, and Transportation will be holding a hearing on security in 5G networks next week. The 5G Supply Chain Security: Threats and Solutions hearing will take place on March 4 at 10 a.m. ET/7 a.m. PT. The committee will hear from experts out of global networking giants Nokia and Ericsson.

"The hearing will examine the security and integrity of the telecommunications supply chain and efforts to secure networks from exploitation in the transition to 5G," the committee page says. It will also look at "the federal government's role in mitigating risks to telecommunications equipment and services in the US and abroad."

https://www.cnet.com/news/us-senate-committee-to-look-at-5g-vendor-security/