A Cult Leader Is Trying to Open a Sex Robot Brothel
In what may be the most Los Angeles story of 2018, if not of all time, the leader of a self-proclaimed feminist cult called Unicult is in the midst of trying to open the "world's first" consent-driven sex robot brothel in West Hollywood.

Unicult's leader, Unicole Unicron, announced their goal of opening Eve's Robot Dreams with an Indiegogo campaign (that recently failed to meet more than one percent of its $155,000 goal). While the concept of a sex robot brothel isn't new—a recent attempt to open the first-ever bot brothel in Houston recently failed after the city council made it illegal to have sex with mechanical objects in local businesses—Eve's Robot Dreams is unique in its focus on consent.

In other words, in order for a client to have sex with one of the brothel's bots, the sex robots would have to agree to it first.

"The goal of Eve’s as an aspect of Unicult is to spread matriarchal change through AI and robots," Unicron told the Daily Beast. Unicron added that Eve's would be staffed entirely by Harmony, a model from Real Doll X that "warms up" to users through conversation. In order to have sex with Harmony, partners have to rack up "conversational points" with Harmony before she "consents" to sex.

https://youtu.be/4X4l7CPZqkU

![](https://hips.hearstapps.com/hmg-prod.s3.amazonaws.com/images/sex-robot-brothel-sq-1543268571.jpg?resize=768:*)