Canadian City Takes Legal Action to Close Sex Doll Brothel
A sex doll rental business located in the Canadian city of Mississauga is facing two charges related to “operating an unlicensed adult business,” according to reports from The Star.

Mississauga’s bylaws only allow nine adult entertainment businesses to operate in the city at the same time. Since there are already this many adult-related businesses in the area, the city argues Aura Dolls is no longer eligible to operate there.

The sex doll brothel was also accused of not having a proper certificate of occupancy. Violators of these provisions can face up to $50,000 in fines and one-year imprisonment.

According to Insauga, in 2018 the gynoid brothel Aura Dolls opened in Toronto, on Yonge Street. But it was forced to move because of a bylaw that banned adult businesses in the area. Therefore, the lease of its intended location was canceled.

The first court appearance was scheduled for Jan. 30. However, at the time of publishing, Aura Dolls had not responded to a message from Future of Sex inquiring about the legal proceedings or whether the business is still in operation.