Inside China’s controversial mission to reinvent the internet
	Please use the sharing tools found via the share button at the top or side of articles. Copying articles to share with others is a breach of FT.com T&Cs and Copyright Policy. Email licensing@ft.com to buy additional rights. Subscribers may share up to 10 or 20 articles per month using the gift article service. More information can be found at https://www.ft.com/tour.
	https://www.ft.com/content/ba94c2bc-6e27-11ea-9bca-bf503995cd6f

	On a cool day late last September, half a dozen Chinese engineers walked into a conference room in the heart of Geneva’s UN district with a radical idea. They had one hour to persuade delegates from more than 40 countries of their vision: an alternative form of the internet, to replace the technological architecture that has underpinned the web for half a century.

Whereas today’s internet is owned by everyone and no one, they were in the process of building something very different — a new infrastructure that could put power back in the hands of nation states, instead of individuals.

The team who had masterminded the New IP (internet protocol) proposal was from the Chinese telecoms giant Huawei, which had sent the largest delegation of any company to the September meeting.

At the gathering, held at the International Telecommunications Union, a UN agency that establishes common global standards for technologies, they presented a simple PowerPoint. It didn’t bother with much detail on how this new network would work, or what specific problem it was solving. Instead, it was peppered with images of futuristic technologies, from life-size holograms to self-driving cars.

The idea was to illustrate that the current internet is a relic that has reached the limits of its technical prowess. It was time, Huawei proposed, for a new global network with a top-down design, and the Chinese should be the ones to build it.