Overhead scanners screen bystanders for signs of COVID-19
To triage a diagnostic shortfall, some hospitals are turning to novel diagnostic technology.



Hospitals are turning to artificial intelligence and machine vision to help with the spread of COVID-19 in novel applications that feel straight out of science fiction. A company called care.ai is working with hospitals to screen visitors by analyzing facial attributes such as sweating and discoloration as well as data from a thermal scan.

This allows hospitals to screen visitors for fever and help reduce the spread of COVID-19 within the hospital. Tampa General Hospital installed the technology to the remaining six entrances of the facility that are still open.

The technology raises tantalizing possibilities, but could ring alarm bells about how its being rolled out. There are also questions about market adoption. Much of the response to COVID-19 infections has been ad hoc, pointing to a lack of vital preparation across a variety of sectors, healthcare foremost among them. Is it feasible to bolt on new technology systems as bandaids for deeper and longstanding structural issues? 

I reached out to Chakri Toleti, founder and CEO of care.ai, to learn more about the facial recognition technology and its place in the escalating crisis.