US finally has plans for the pandemic it now leads in infections
Meanwhile, epidemiologists facing pushback for doing science.

On Thursday, the US passed a grim milestone: it surpassed China as the country with the largest number of confirmed Coronavirus cases. The milestone was reached during what appears to be a growing public tug-of-war between senior Trump administration officials, who wish to see restrictions lifted as quickly as possible, and public health experts, who argue they're clearly still required for the time being. That tension may end up playing out in the implementation of a new plan being developed to guide states through their response to the pandemic.

One casualty in this fight: the work of epidemiologists. As these researchers continue to test the impact of different restrictions on the spread of infections, their models are necessarily producing different numbers. Those differences are now being dragged into the intensely political arguments about how best to respond to the pandemic.

Grim numbers

As of this writing, the world has seen slightly under 560,000 confirmed cases of SARS-CoV-2, according to the Johns Hopkins Coronavirus Resource Center. Of those, 86,000 (about 15 percent) are in the US. China, where the pandemic originated, is near 82,000 cases, followed by Italy with a bit over 80,000. All of the other countries with over 10,000 confirmed cases are European, with the exception of Iran.

China is indicating that the majority of its cases are now due to secondary re-introduction: the virus has largely stopped circulating within the populace, but travelers from countries with unchecked spread are now bringing it back with them.

China is in a position to track new infections because it imposed significant restrictions on its population and then ramped up testing so that new instances of transmission could be identified quickly. Despite having advanced warning of the spread of SARS-CoV-2, the US's failure to ramp up testing allowed the virus to establish itself in many areas of the country. As a result, many states have been compelled to place severe restrictions on their populace in the hope of limiting the rapidly spreading infections.