Facebook Awards $55,000 for Flaw That Could Lead to Account Hijacking
A researcher has earned $55,000 from Facebook for reporting a serious vulnerability that could have been exploited by hackers to steal access tokens and hijack accounts.

India-based researcher Amol Baikar discovered in December that the “Login with Facebook” feature, which allows users to log in to other websites with their Facebook account, was affected by a vulnerability.

Baikar found that an attacker could hijack the OAuth flow — the feature uses the OAuth 2.0 authorization protocol — and steal a user’s access tokens for apps that use the OAuth flow, such as Instagram and Oculus. For the attack to work, a hacker would have needed to convince the targeted user to visit a malicious website.

Facebook was informed about the vulnerability on December 16 and a patch was released within a week. However, the researcher found a way to bypass the patch and a more efficient fix was rolled out in mid-January.

Baikar said he received a total of $55,000 from Facebook for his initial report and the patch bypass. The researcher said this was the highest bounty paid out by the social media giant for a client-side account takeover vulnerability.