Agencies Warn of Coronavirus-Related Scams
The government is trying to curb scammers preying on the public’s fears.

Several federal agencies issued warnings this week alert the public to look out for potential scams and criminal activity related to the new coronavirus.

“Scammers are taking advantage of fears surrounding the Coronavirus,” the Federal Trade Commission said in an alert. “They’re setting up websites to sell bogus products, and using fake emails, texts, and social media posts as a ruse to take your money and get your personal information.”

FTC warns scammers may appear to promote awareness and virus prevention tips or ask you to donate to relief efforts. Perpetrators often email potential victims claiming to be from the Centers for Disease Control and Prevention or other health organizations in hopes of getting victims to click on links that install malware on their computers. The Cybersecurity and Infrastructure Security Agency also warns the public to be wary of websites that ask for personnel information or contain attachments, hyperlinks or social media pleas, texts or calls related to COVID-19.