WELCOME TO PWN2OWN 2020 - THE SCHEDULE AND LIVE RESULTS
Welcome to Pwn2Own 2020! This year’s contest will be the first where all attempts occur remotely. We have contestants from around the world ready to demonstrate some amazing research, and we have ZDI researchers ready to run and verify their attempts. Our returning partner Microsoft and sponsor VMware will also be online with us, as well as other affected vendors. We might not be in the same room, but we’re virtually together and ready for a fantastic contest. 

The schedule for today is posted below, and we’ll be updating this blog throughout the day with results and updates. We’ll post a full summary of today’s events (including videos of the attempts) tomorrow morning.

As always, we started the contest with a random drawing to determine the order of attempts. We have four attempts scheduled for today and four queued up for tomorrow. The full schedule for Day One is below (all times Pacific [UTC -7:00]). We will update this schedule with results as they become available.