PWN2OWN DAY TWO – RESULTS AND MASTER OF PWN
The final day of Pwn2Own 2020 came to a close yesterday, but not before a bit of drama, intrigue, and more great research. We saw two successful attempts, which resulted in four new bugs earning $90,000. That brings the two-day total to $270,000. Here’s a quick recap of Day Two and the entire event: 

https://youtu.be/u1udr7j9MQA