Indian state of Tamil Nadu announces blockchain infrastructure plan
The Indian state of Tamil Nadu unveils new technology projects, including a blockchain infrastructure, to create easier government processes.

Tamil Nadu chief minister Edapaddi K. Palaniswami announced Tamil Nadu's adoption of an ethical blockchain policy, according to The New Indian Express. 

Palaniswami also announced that Tamil Nadu intends to launch a project to provide Knowledge Proof Identity-based services delivery, which, according to the minister, will revolutionize the way the government provides its services to citizens.

The project will be implemented by establishing and leveraging the State Family Database (SFDB) and Blockchain Backbone infrastructure. The blockchain infrastructure functions as a single source of truth for all government data and processes. It may also be used to reimagine governance and encourage blockchain use in the state. 

By launching the Blockchain Backbone infrastructure, the state also acknowledges the need to create a forum to promote collaboration between industry, start-ups, and academia to build an ecosystem. 

The state of Tamil Nadu has also reportedly adopted an ethical artificial intelligence policy that aims to take advantage of the advantages of AI while keeping it safe and ethically compatible with human values. It also adopted a policy similar to cybersecurity to protect information and combat cybercrimes.