America might finally get open banking — but not without a fight
Proposed rulemaking on Dodd-Frank Section 1033 could fundamentally reshape American finance.

Late last week, the Consumer Financial Protection Bureau quietly published a statement that could fundamentally change the fintech industry.

The catchily-named "Advance Notice of Proposed Rulemaking on Consumer Access to Financial Records" (colloquially known as the ANPR) explained that the CFPB plans to create new regulation to implement Section 1033 of the Dodd-Frank Act, and is seeking stakeholder input before it does so. Yes, that's a lot of jargon. But look past that, and the consequences are significant: The future of fintech is now on the line.

The Dodd-Frank Act is a behemoth of financial legislation, passed in the aftermath of the last financial crisis. It sought to overhaul the system and revive trust in a sector where that was deeply lacking. In doing so, it also laid the groundwork for the following decade's fintech boom. That's partly down to the unassuming Section 1033. "It says consumers have the right to access their financial account information in a machine-readable format," Plaid's global head of policy, John Pitts, told Protocol. That allows companies like Mint to aggregate your accounts, for instance, and lets cash-flow based lenders access your bank details so they can effectively underwrite loans. It's the closest thing the U.S. has to something like the U.K.'s Open Banking regime, or the EU's Payment Services Directive 2. Section 1033, Pitts said, "underpins all of fintech."

https://www.protocol.com/open-banking-dodd-frank-data-fintech