The Week in Ransomware - June 11th 2021 - Under Pressure
https://www.bleepingcomputer.com/news/security/the-week-in-ransomware-june-11th-2021-under-pressure/

It has been quite the week when it comes to ransomware, with ransoms being paid, ransoms being taken back, and a ransomware gang shutting down.

This week's biggest news was the FBI announcing that they were able to recover the majority of the $4.4 million ransom payment paid by Colonial Pipeline. It is not entirely clear how they obtained the private key for the cryptocurrency wallet, but it is believed DarkSide stored it on a seized server.